PDFBOX-1939: Store all stroke information in the graphics state
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1571581 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/util/operator/SetLineMiterLimit.java
@@ -44,6 +44,6 @@ public class SetLineMiterLimit extends OperatorProcessor
     public void process(PDFOperator operator, List<COSBase> arguments) throws IOException
     {
         COSNumber miterLimit = (COSNumber)arguments.get( 0 );
-        context.getGraphicsState().setMiterLimit( miterLimit.doubleValue() );
+        context.getGraphicsState().setMiterLimit( miterLimit.floatValue() );
     }
 }
